+++ DotNet/Normalization/Program.cs
@@ -4,6 +4,7 @@ using HealthChecks.UI.Client;
 using LantanaGroup.Link.Normalization.Application.Interfaces;
 using LantanaGroup.Link.Normalization.Application.Managers;
 using LantanaGroup.Link.Normalization.Application.Models.Messages;
+using LantanaGroup.Link.Normalization.Application.Operations;
 using LantanaGroup.Link.Normalization.Application.Serializers;
 using LantanaGroup.Link.Normalization.Application.Services;
 using LantanaGroup.Link.Normalization.Application.Settings;
@@ -221,6 +222,8 @@ static void RegisterServices(WebApplicationBuilder builder)
     builder.Services.AddTransient<RetryJob>();
 
     builder.Services.AddSingleton<IConditionalTransformationEvaluationService, ConditionalTransformationEvaluationService>();
+
+    builder.Services.AddHostedService<CopyPropertyOperationService>();
 
     if (consumerSettings != null && !consumerSettings.DisableConsumer)
     {
